Refactor docker buildx and using a Dockerfile template
The installation steps to install buildx as dedicated tool is removed and we use the native docker buildx command
instead. Generate the Dockerfile from a template instead of duplicating the arguments and parameters with docker build
args. Use the CircleCI matrix feature to build all architectures for every feature or bug fix branch and not just for
amd64 architecture. We provide OCI images for each architecture as build artifact in CircleCI and provide a README with
build and usage instructions.